requirements/requirements.txt
# python3 -m pip install -r requirements/requirements.txt
# sudo python3 -m pip install -r requirements/requirements.txt
python-dotenv

websocket-client # import websocket
rel

httpx

Flask
flask_caching
flask-sock
Flask-SocketIO
flask-cors

uwsgi
gunicorn

# breaks older python installs. Is this needed for docker?
uWSGI==2.0.17.1

pycoingecko

# https://pypi.org/project/py-kvstore/ / https://github.com/Reecepbcups/py_kvstore
py-kvstore==0.0.2